Remove "-0" from Cafe and Celeste Tech Train
The number at the end of an ID after a single dash indicates the version of that cart on lexaloffle. If you load a cart with this number in the ID, it will always download the one particular version of the cart it refers to (lexaloffle keeps archives of old versions). In order to load the *latest* version of the cart, you should just enter the ID without the number. Example: #noeleste-0 will load the OG noeleste, while #noeleste will load noelle, the 2.1 version with changed levels.