Added cchristmas mods +Athos -Squiggle
Linked to mods only available on the calendar with [github](link), but I tried to use lexaloffle links where I know they existed (might have missed some). I linked to site.p8 as a download, since a version came out through a discord post that is not on the calendar. Not sure if linking to a discord download is good practice anymore or if I did it in the best way. I did not include squiggle (or spiral/unknown/S/S+ that have come out since the last update for that matter) because I feel like the difficulty generally makes them unviable for a casual mod list (have a similar gripe with jank but never really want to make a huge deal about deleting a mod or whatever, especially since I don't think other people agree with me a lot there). Did this all at once and felt kind of tired while doing it so it would probably be good if somebody checked my work for typoes/general brain farts. For the "Latest" section I got rid of everything pre-2023 and just added the cchristmas mods as one link since the calendar should still work and contain everything (might make it hard to find sugar rush though if somebody's using the calendar for stuff like that). I don't know how smellyfishtiks wants their name capitalized but I'm also not sure if anybody's staring too hard at the capitalization to begin with. As always mod summaries are just kind of whatever felt right at the moment with no format or standardization, I did not pull from the calendar since a lot of them felt too long/a bit weird to have in the summary section