Added moty stuff in descriptions
For mods that won added a short blurb saying they won an award with the capitalization used in the awards form. Since no mod won two awards I think this works for right now, although I'd be super down to change it if we deem that better. Also, wasn't super sure how to make it clear each award was for a mod so stuff like "won Best Speedgame 2023" might be a bit weird for somebody who thinks it means more than just celeste mod speedgames. Not sure how to fix though