Condense ID and Link columns
Having less columns looks better. And since IDs can only exist if there
is a Lexaloffle upload, no new information is conveyed by specifying
"Lexaloffle" in a different column.

I can imagine that copying text that is also a link can be difficult on
mobile, but there is no reason to copy cart IDs on mobile. On PC just
triple click the cell to select the ID and then copy.